requirements.txt
django
django-tables2
django-filter
django-bootstrap3